chore(#3351): satisfy giant-file ratchet — move reclaim helper to tmux_watcher/placeholder_reclaim.rs
Move the reclaim helper + r21 record-drop wrapper into a new child module
(S1 pattern), condense wiring comments, and raise the two root baselines
to the measured post-move values with rationale (tmux_watcher +101 wiring
— S2 re-lowers; inflight +58 compare-and-clear pinned to file-private
persistence internals).
